requirements.txt
# Unified requirements for Kavach Monorepo (UI + 2 Backends)
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
pydantic>=2.9.0
pydantic-settings>=2.6.0
cryptography>=43.0.0
lxml>=5.3.0
signxml>=4.0.0
python-multipart>=0.0.12
httpx>=0.27.0
redis>=5.0.0
pyngrok>=7.2.0
python-dotenv>=1.0.0
